Support no_std; Move all tests to the tests dir
